Compare the parameter index of the thesis MC in one width
The clean rebuild in the C locale reported the last warning of the tree:
the numerator loop of replaceParameter compared a signed parameter index
with an unsigned position, where the denominator loop beside it already
cast. Both compare as std::size_t now, and the tree builds with no
warnings.

+++ src/core/loopshaping/algorithm_mc_thesis.cpp
@@ -94,7 +94,7 @@ std::unique_ptr<LtiSystem> AlgorithmMcThesis::replaceParameter(LtiSystem * box, 
     numerator.reserve(box->numerator().size());
     for (std::size_t j = 0; j < box->numerator().size(); ++j) {
         Parameter & old = box->numerator()[j];
-        numerator.push_back(parameter == j + 1
+        numerator.push_back(static_cast<std::size_t>(parameter) == j + 1
                 ? Parameter(old.name(), range, range.min)
                 : old);
     }
